pyproject.toml
[tool.poetry]
name = "hydrazenlauncher"
version = "0.1.0"
description = ""
authors = ["Carel van Niekerk <cvanniekerk@hhu.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
hydra-core = "^1.3.2"
hydra-zen = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
